chore: remove storybook; improve deps
