+++ babel.config.js
@@ -14,10 +14,9 @@ module.exports = {
   ],
 
   plugins: (process.env.NODE_ENV === 'development'
-    ? ['react-hot-loader/babel', 'react-refresh/babel']
+    ? [ 'react-refresh/babel']
     : []
   ).concat([
-    '@emotion',
     '@babel/plugin-syntax-dynamic-import',
     '@babel/plugin-transform-runtime',
     ['@babel/plugin-proposal-decorators', { legacy: true }],
