+++ client/main.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Root from './config/root'
+
+import './assets/scss/main.scss'
+
+const target = document.getElementById('root')
+
+ReactDOM.render(<Root />, target)
