+++ package.json
@@ -13,7 +13,7 @@
         "start": "cross-env NODE_ENV=production && node start.js",
         "clean": "rimraf dist",
         "lint": "eslint ./client/**/*.js ./server/**/*.js ",
-        "prettier": "prettier './client/**/*.js' './server/**/*.js' --write ",
+        "prettier": "prettier './client/**/*.js' './client/**/*.jsx' './server/**/*.js' --write ",
         "lint:server": "eslint ./server/**/**.js",
         "test": "jest",
         "test:watch": "jest --watch",
@@ -48,8 +48,8 @@
         }
     },
     "dependencies": {
-        "@babel/cli": "^7.14.5",
-        "@babel/core": "^7.14.6",
+        "@babel/cli": "^7.14.8",
+        "@babel/core": "^7.15.0",
         "@babel/plugin-proposal-class-properties": "^7.14.5",
         "@babel/plugin-proposal-decorators": "^7.14.5",
         "@babel/plugin-proposal-export-default-from": "^7.14.5",
@@ -60,43 +60,39 @@
         "@babel/plugin-proposal-throw-expressions": "^7.14.5",
         "@babel/plugin-syntax-dynamic-import": "^7.8.3",
         "@babel/plugin-syntax-import-meta": "^7.10.4",
-        "@babel/plugin-transform-runtime": "^7.14.5",
-        "@babel/preset-env": "^7.14.7",
+        "@babel/plugin-transform-runtime": "^7.15.0",
+        "@babel/preset-env": "^7.15.0",
         "@babel/preset-flow": "^7.14.5",
         "@babel/preset-react": "^7.14.5",
         "@babel/preset-stage-3": "^7.8.3",
-        "@babel/preset-typescript": "^7.14.5",
+        "@babel/preset-typescript": "^7.15.0",
         "@babel/register": "^7.14.5",
-        "@babel/runtime": "^7.14.6",
+        "@babel/runtime": "^7.14.8",
         "@emotion/babel-plugin": "^11.3.0",
         "babel-core": "7.0.0-bridge.0",
         "babel-eslint": "10.1.0",
-        "babel-jest": "^27.0.5",
+        "babel-jest": "^27.0.6",
         "babel-loader": "^8.2.2",
         "babel-plugin-css-modules-transform": "^1.6.2",
-        "babel-plugin-emotion": "^11.0.0",
         "babel-polyfill": "^6.26.0",
-        "babel-preset-react-boilerplate": "1.1.1",
-        "babel-preset-react-hmre": "1.1.1",
         "classnames": "^2.3.1",
         "colors": "^1.4.0",
         "connected-react-router": "^6.9.1",
         "cookie-parser": "^1.4.5",
-        "core-js": "^3.15.1",
+        "core-js": "^3.16.0",
         "cors": "^2.8.5",
         "cross-env": "^7.0.3",
         "dotenv": "^10.0.0",
         "express": "^4.17.1",
         "history": "^4.10.1",
         "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "17.0.1",
+        "react-dom": "17.0.1",
         "react-helmet": "^6.1.0",
-        "react-hot-loader": "^4.13.0",
         "react-redux": "^7.2.4",
         "react-refresh": "^0.10.0",
         "react-router-dom": "^5.2.0",
-        "redux": "^4.1.0",
+        "redux": "^4.1.1",
         "redux-devtools-extension": "^2.13.9",
         "redux-thunk": "^2.3.0",
         "sockjs": "^0.3.21",
@@ -107,21 +103,13 @@
     "devDependencies": {
         "@hot-loader/react-dom": "^17.0.1",
         "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
-        "@storybook/addon-actions": "^6.3.0",
-        "@storybook/addon-info": "^5.3.21",
-        "@storybook/addon-knobs": "^6.2.9",
-        "@storybook/addon-links": "^6.3.0",
-        "@storybook/addon-viewport": "^6.3.0",
-        "@storybook/addons": "^6.3.0",
-        "@storybook/cli": "^6.3.0",
-        "@storybook/react": "^6.3.0",
-        "autoprefixer": "^10.2.6",
-        "copy-webpack-plugin": "^9.0.0",
-        "css-loader": "5.2.6",
-        "css-minimizer-webpack-plugin": "^3.0.1",
-        "cssnano": "^5.0.6",
+        "autoprefixer": "^10.3.1",
+        "copy-webpack-plugin": "^9.0.1",
+        "css-loader": "6.2.0",
+        "css-minimizer-webpack-plugin": "^3.0.2",
+        "cssnano": "^5.0.7",
         "emotion-theming": "^11.0.0",
-        "eslint": "^7.29.0",
+        "eslint": "^7.32.0",
         "eslint-config-airbnb": "18.2.1",
         "eslint-config-prettier": "^8.3.0",
         "eslint-loader": "4.0.2",
@@ -135,31 +123,29 @@
         "glob": "^7.1.7",
         "hard-source-webpack-plugin": "^0.13.1",
         "html-webpack-plugin": "5.3.2",
-        "husky": "6.0.0",
+        "husky": "7.0.1",
         "image-webpack-loader": "^7.0.1",
-        "jest": "^27.0.5",
-        "jest-cli": "^27.0.5",
-        "mini-css-extract-plugin": "^1.6.0",
+        "jest": "^27.0.6",
+        "jest-cli": "^27.0.6",
+        "mini-css-extract-plugin": "^2.2.0",
         "node-hot-loader": "^1.21.4",
-        "node-sass": "^6.0.0",
-        "nodemon": "^2.0.7",
-        "optimize-css-assets-webpack-plugin": "^6.0.1",
-        "postcss": "^8.3.5",
+        "node-sass": "^6.0.1",
+        "nodemon": "^2.0.12",
+        "postcss": "^8.3.6",
         "postcss-import": "^14.0.2",
-        "postcss-loader": "^6.1.0",
+        "postcss-loader": "^6.1.1",
         "postcss-preset-env": "^6.7.0",
-        "prettier": "^2.3.1",
+        "prettier": "^2.3.2",
         "raw-loader": "^4.0.2",
-        "react-addons-test-utils": "^15.6.2",
-        "regenerator-runtime": "^0.13.7",
+        "regenerator-runtime": "^0.13.9",
         "rimraf": "3.0.2",
         "sass-loader": "12.1.0",
-        "style-loader": "2.0.0",
-        "tailwindcss": "^2.2.2",
-        "terser-webpack-plugin": "^5.1.3",
+        "style-loader": "3.2.1",
+        "tailwindcss": "^2.2.7",
+        "terser-webpack-plugin": "^5.1.4",
         "thread-loader": "^3.0.4",
         "url-loader": "4.1.1",
-        "webpack": "^5.40.0",
+        "webpack": "^5.49.0",
         "webpack-cli": "^4.7.2",
         "webpack-dev-server": "^3.11.2",
         "webpack-node-externals": "^3.0.0",
@@ -172,8 +158,8 @@
     },
     "husky": {
         "hooks": {
-            "pre-commit": "cross-env npm run lint && npm test",
-            "pre-push": "cross-env npm run lint && npm test"
+            "pre-commit": "cross-env yarn run lint && yarn test",
+            "pre-push": "cross-env yarn run lint && yarn test"
         }
     }
 }
