+++ server/.babelrc
@@ -8,9 +8,8 @@
       "@babel/react",
       "@babel/typescript"
     ],
-  
+
     "plugins": [
-      "react-hot-loader/babel",
       "@babel/plugin-proposal-export-default-from",
       "@babel/plugin-proposal-class-properties",
       ["@babel/plugin-proposal-decorators",{ "legacy": true }]
