--- stories/0-Welcome.stories.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { linkTo } from '@storybook/addon-links';
-import { Welcome } from '@storybook/react/demo';
-
-export default {
-  title: 'Welcome',
-};
-
-export const toStorybook = () => <Welcome showApp={linkTo('Button')} />;
-
-toStorybook.story = {
-  name: 'to Storybook',
-};
